feat: Adds AHB repository for data access
Implements the AHB repository to handle data access logic
for Anwendungshandbuch (AHB) data. It includes search functionality
with filtering, sorting, and pagination, as well as retrieval
of complete AHB data in JSON or XLSX format. The repository uses
TypeORM to interact with the database and includes error handling
for not found scenarios. It also contains mapping functions to
transform database entities into API models.