docs(ahb.db): Update Documentation How To Update Database (#633)
* Updates AHB database documentation

Updates the README file to reflect the current process of creating the AHB database.

It now references the `fundament` python package and the corresponding script for database creation, as well as the private repository containing the XML files.

* Adds documentation for the AHB database

Provides a comprehensive README for the `ahb.db` database,
explaining its purpose, structure, generation process,
and maintenance.

The documentation clarifies the database's role in
the AHB Tabellen application, its dependency on
the private `xml-migs-and-ahbs` repository, and
the steps required to update the database with new data.

* Documents database update process and location

Adds documentation on how to update the AHB database,
including its source repository location and why it's private.
The README in `src/server/data` now points to the root README for database information,
and the detailed database documentation has been moved to the root README.

* prettier readme

* Revise README for AHB database management

Updated instructions for database generation and maintenance, including links to private repository and release process.

* Revise AHB data update instructions in README

Updated instructions for updating the database with new AHB data, including manual XML file downloads and CI error handling.

---------

Co-authored-by: konstantin <konstantin.klein@hochfrequenz.de>